requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.11
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
aiosqlite==0.21.0
    # via -r requirements.in
annotated-types==0.7.0
    # via
    #   -r requirements.in
    #   pydantic
attrs==25.4.0
    # via aiohttp
bidict==0.23.1
    # via
    #   -r requirements.in
    #   python-socketio
blinker==1.9.0
    # via
    #   -r requirements.in
    #   flask
certifi==2025.11.12
    # via
    #   -r requirements.in
    #   requests
cffi==2.0.0
    # via
    #   -r requirements.in
    #   cryptography
charset-normalizer==3.4.4
    # via
    #   -r requirements.in
    #   requests
click==8.3.1
    # via
    #   -r requirements.in
    #   flask
cryptg==0.5.2
    # via -r requirements.in
cryptography==46.0.3
    # via -r requirements.in
deprecated==1.3.1
    # via
    #   -r requirements.in
    #   limits
docker==7.1.0
    # via -r requirements.in
filelock==3.20.0
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   torch
    #   transformers
flask==3.1.2
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-limiter
    #   flask-socketio
flask-cors==4.0.0
    # via -r requirements.in
flask-limiter==4.0.0
    # via -r requirements.in
flask-socketio==5.5.1
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.10.0
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via
    #   -r requirements.in
    #   wsproto
hf-xet==1.2.0
    # via
    #   -r requirements.in
    #   huggingface-hub
huggingface-hub==0.36.0
    # via
    #   -r requirements.in
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   -r requirements.in
    #   requests
    #   yarl
iniconfig==2.3.0
    # via
    #   -r requirements.in
    #   pytest
itsdangerous==2.2.0
    # via
    #   -r requirements.in
    #   flask
jinja2==3.1.6
    # via
    #   -r requirements.in
    #   flask
    #   torch
joblib==1.5.2
    # via
    #   -r requirements.in
    #   scikit-learn
limits==5.6.0
    # via
    #   -r requirements.in
    #   flask-limiter
markdown-it-py==4.0.0
    # via
    #   -r requirements.in
    #   rich
markupsafe==3.0.3
    # via
    #   -r requirements.in
    #   flask
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via
    #   -r requirements.in
    #   markdown-it-py
mpmath==1.3.0
    # via
    #   -r requirements.in
    #   sympy
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
networkx==3.6
    # via
    #   -r requirements.in
    #   torch
numpy==2.2.6
    # via
    #   -r requirements.in
    #   scikit-learn
    #   scipy
    #   transformers
ordered-set==4.1.0
    # via
    #   -r requirements.in
    #   flask-limiter
packaging==25.0
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   limits
    #   pytest
    #   transformers
pillow==12.0.0
    # via
    #   -r requirements.in
    #   sentence-transformers
pluggy==1.6.0
    # via
    #   -r requirements.in
    #   pytest
prometheus-client==0.23.1
    # via -r requirements.in
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
psutil==7.1.3
    # via -r requirements.in
pyaes==1.6.1
    # via
    #   -r requirements.in
    #   telethon
pyasn1==0.6.1
    # via
    #   -r requirements.in
    #   rsa
pycparser==2.23
    # via
    #   -r requirements.in
    #   cffi
pydantic==2.12.4
    # via -r requirements.in
pydantic-core==2.41.5
    # via
    #   -r requirements.in
    #   pydantic
pygments==2.19.2
    # via
    #   -r requirements.in
    #   pytest
    #   rich
pytest==9.0.1
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-mock
pytest-asyncio==1.3.0
    # via -r requirements.in
pytest-mock==3.15.1
    # via -r requirements.in
python-engineio==4.12.3
    # via
    #   -r requirements.in
    #   python-socketio
python-socketio==5.15.0
    # via
    #   -r requirements.in
    #   flask-socketio
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
rapidfuzz==3.9.6
    # via -r requirements.in
redis==7.1.0
    # via -r requirements.in
regex==2025.11.3
    # via
    #   -r requirements.in
    #   transformers
requests==2.32.5
    # via
    #   -r requirements.in
    #   docker
    #   huggingface-hub
    #   transformers
rich==14.2.0
    # via
    #   -r requirements.in
    #   flask-limiter
rsa==4.9.1
    # via
    #   -r requirements.in
    #   telethon
safetensors==0.7.0
    # via
    #   -r requirements.in
    #   transformers
scikit-learn==1.7.2
    # via
    #   -r requirements.in
    #   sentence-transformers
scipy==1.16.3
    # via
    #   -r requirements.in
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.1.2
    # via -r requirements.in
simple-websocket==1.1.0
    # via
    #   -r requirements.in
    #   python-engineio
sqlalchemy==2.0.44
    # via -r requirements.in
sympy==1.14.0
    # via
    #   -r requirements.in
    #   torch
telethon==1.42.0
    # via -r requirements.in
tgcrypto==1.2.5
    # via -r requirements.in
threadpoolctl==3.6.0
    # via
    #   -r requirements.in
    #   scikit-learn
tokenizers==0.22.1
    # via
    #   -r requirements.in
    #   transformers
torch==2.9.1
    # via
    #   -r requirements.in
    #   sentence-transformers
tqdm==4.67.1
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.57.1
    # via
    #   -r requirements.in
    #   sentence-transformers
typing-extensions==4.15.0
    # via
    #   -r requirements.in
    #   aiosignal
    #   aiosqlite
    #   flask-limiter
    #   huggingface-hub
    #   limits
    #   pydantic
    #   pydantic-core
    #   pytest-asyncio
    #   sentence-transformers
    #   sqlalchemy
    #   torch
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   -r requirements.in
    #   pydantic
urllib3==2.5.0
    # via
    #   -r requirements.in
    #   docker
    #   requests
werkzeug==3.1.3
    # via
    #   -r requirements.in
    #   flask
wrapt==2.0.1
    # via
    #   -r requirements.in
    #   deprecated
wsproto==1.3.2
    # via
    #   -r requirements.in
    #   simple-websocket
yarl==1.22.0
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
